refactor(query): deepen command semantics seam
